core: move handlers to new folder
