+++ BlasII.Randomizer/Shuffle/Pools.cs
@@ -1,5 +1,4 @@
-﻿using BlasII.Randomizer.Items;
-using BlasII.Randomizer.Models;
+﻿using BlasII.Randomizer.Models;
 using System;
 using System.Collections;
 using System.Collections.Generic;
